+++ packages/mangrove-solidity/src/strategies/interfaces/IMakerLogic.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier:	BSD-2-Clause
 
-// AdvancedCompoundRetail.sol
+// IMakerLogic.sol
 
 // Copyright (c) 2021 Giry SAS. All rights reserved.
 
@@ -9,18 +9,20 @@
 // 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
 // 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-pragma solidity ^0.8.10;
-pragma abicoder v2;
-import "mgv_src/strategies/multi_user/abstract/Persistent.sol";
-import "mgv_src/strategies/routers/SimpleRouter.sol";
 
-contract OasisLike is MultiUserPersistent {
-  constructor(IMangrove _MGV, address deployer)
-    MultiUserPersistent(_MGV, new SimpleRouter(), 30_000)
-  {
-    if (deployer != msg.sender) {
-      set_admin(deployer);
-      router().set_admin(deployer);
-    }
-  }
+pragma solidity >=0.8.0;
+pragma abicoder v2;
+import "./IOfferLogic.sol";
+
+interface IMakerLogic is IOfferLogic {
+  
+  function newOffer(
+    IERC20 outbound_tkn, // address of the ERC20 contract managing outbound tokens
+    IERC20 inbound_tkn, // address of the ERC20 contract managing outbound tokens
+    uint wants, // amount of `inbound_tkn` required for full delivery
+    uint gives, // max amount of `outbound_tkn` promised by the offer
+    uint gasreq, // max gas required by the offer when called. If maxUint256 is used here, default `ofr_gasreq` will be considered instead
+    uint gasprice, // gasprice that should be consider to compute the bounty (Mangrove's gasprice will be used if this value is lower)
+    uint pivotId
+  ) external payable returns (uint);
 }