+++ packages/mangrove-solidity/src/strategies/offer_forwarder/OfferForwarder.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier:	BSD-2-Clause
 
-// AdvancedAaveRetail.sol
+// OfferForwarder.sol
 
 // Copyright (c) 2021 Giry SAS. All rights reserved.
 
@@ -11,31 +11,46 @@
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 pragma solidity ^0.8.10;
 pragma abicoder v2;
-import "mgv_src/strategies/single_user/abstract/Persistent.sol";
-import "mgv_src/strategies/routers/AaveDeepRouter.sol";
+import "mgv_src/strategies/offer_forwarder/abstract/PersistentForwarder.sol";
+import "mgv_src/strategies/interfaces/IMakerLogic.sol";
+import "mgv_src/strategies/routers/SimpleRouter.sol";
 
-contract AdvancedAaveRetail is Persistent {
-  constructor(
-    IMangrove _mgv,
-    address _addressesProvider,
-    address deployer
-  ) Persistent(_mgv, 30_000, new AaveDeepRouter(_addressesProvider, 0, 2)) {
-    // Router reserve is by default `router.address`
-    // use `set_reserve(addr)` to change this
-    router().set_admin(deployer);
+contract OfferForwarder is IMakerLogic, PersistentForwarder {
+  constructor(IMangrove _MGV, address deployer)
+    PersistentForwarder(_MGV, new SimpleRouter())
+  {
+    set_gasreq(50_000);
     if (deployer != msg.sender) {
       set_admin(deployer);
+      router().set_admin(deployer);
     }
   }
-
-  // overriding put to leverage taker's liquidity on aave
-  // this function will deposit incoming liquidity to increase borrow power during trade
-  function __put__(uint amount, ML.SingleOrder calldata order)
-    internal
-    override
-    returns (uint missingPut)
+  
+  // As imposed by IMakerLogic we provide an implementation of newOffer for this contract
+  function newOffer(
+    IERC20 outbound_tkn,
+    IERC20 inbound_tkn,
+    uint wants,
+    uint gives,
+    uint gasreq,
+    uint gasprice,
+    uint pivotId)
+    external
+    payable
+    returns (uint offerId)
   {
-    push(IERC20(order.inbound_tkn), amount);
-    return 0;
+    offerId = _newOffer(
+      NewOfferData({
+        outbound_tkn: outbound_tkn,
+        inbound_tkn: inbound_tkn,
+        wants: wants,
+        gives: gives,
+        gasreq: gasreq,
+        gasprice: gasprice,
+        pivotId: pivotId,
+        caller: msg.sender,
+        fund: msg.value
+        })
+    );
   }
 }