+++ packages/mangrove-solidity/test/periphery/MgvOrder.t.sol
@@ -247,7 +247,7 @@ contract MangroveOrder_Test is MangroveTest {
       gasForMarketOrder: 6_500_000,
       blocksToLiveForRestingOrder: 0 //NA
     });
-    vm.expectRevert("MultiUser/derive_gasprice/NotEnoughProvision");
+    vm.expectRevert("Forwarder/derive_gasprice/NotEnoughProvision");
     mgo.take{value: 0.0001 ether}(buyOrder);
   }